refactor valitation's code
