+++ src/hooks/index.js
@@ -1,1 +1,2 @@
 export { useModal, ModalProvider } from './modal';
+export { useValidation } from './validation';
